fix: change float types to double

+++ SolarWatch/SolarWatch/Services/ICoordAndDateProcessor.cs
@@ -2,6 +2,6 @@
 
 public interface ICoordAndDateProcessor
 {
-    public Task<string> GetSunriseTime(float lat, float lon, string date);
-    public Task<string> GetSunsetTime(float lat, float lon, string date);
+    public Task<string> GetSunriseTime(double lat, double lon, string date);
+    public Task<string> GetSunsetTime(double lat, double lon, string date);
 }
